[update]Ruifeng-to be discussed via vedio call

+++ VenueManagement_VenueList.java
@@ -1,11 +1,13 @@
 
 import java.util.ArrayList;
+import java.util.Calendar;
 
 import org.eclipse.swt.SWT;
 import org.eclipse.swt.events.DisposeEvent;
 import org.eclipse.swt.events.DisposeListener;
 import org.eclipse.swt.widgets.Composite;
 import org.eclipse.swt.widgets.Display;
+import org.eclipse.swt.widgets.MessageBox;
 import org.eclipse.swt.widgets.Shell;
 import org.eclipse.swt.widgets.TableItem;
 import org.eclipse.ui.forms.widgets.FormToolkit;
@@ -21,8 +23,10 @@ public class VenueManagement_VenueList extends Composite {
 
 	private final FormToolkit toolkit = new FormToolkit(Display.getCurrent());
 	private Table venueTable;
+	private String[] venueInfoArray = { "Name", "Location", "Type", "Capacity" };
+    private int[] venueInfoSignatureArray = { MACRO.TEXT, MACRO.VENUELOCATION, MACRO.VENUETYPE, MACRO.INT};
     
-	private ArrayList<Venue> venueList;
+    private Date today = new Date();
 	/**
 	 * Create the composite.
 	 * @param parent
@@ -90,13 +94,18 @@ public class VenueManagement_VenueList extends Composite {
 	    btnEdit.addSelectionListener(new edit());
 	    btnView.addSelectionListener(new viewBookingInfo());
 	    
+		Calendar calendar = Calendar.getInstance(); // today
+		today.setYear(calendar.get(Calendar.YEAR));
+		today.setMonth(calendar.get(Calendar.MONTH)+1);
+		today.setDay(calendar.get(Calendar.DATE));
+	    
 	    importVenueListData();
 
 	}
 	
 	public void importVenueListData() {
 		DatabaseReader db = new DatabaseReader();
-		venueList = db.getVenues();
+		ArrayList<Venue> venueList = db.getVenues();
 		/* update the venue list table */
 		for (int i=0; i<venueList.size(); i++){
 			TableItem temp = new TableItem(venueTable, SWT.None);
@@ -112,7 +121,23 @@ public class VenueManagement_VenueList extends Composite {
 		public void widgetSelected(SelectionEvent e) {
 			venueTable.deselectAll();
 			Shell addVenueShell = new Shell(getDisplay());
-			VenueListAddItem addVenuePage = new VenueListAddItem(addVenueShell, SWT.None, venueTable);
+			Image icon = new Image(getDisplay(), "resources/eManager.png");
+			addVenueShell.setText("eManager - Add a venue");
+			addVenueShell.setImage(icon);
+			AbstractAdd addVenuePage = new AbstractAdd(addVenueShell, SWT.None, venueInfoArray, venueInfoSignatureArray, venueTable){
+				// submit info
+				public void onSubmit() {
+					// update database
+					String[] venueInfoList = getStringList();
+					Venue venue = new Venue(venueInfoList[0], venueInfoList[1], venueInfoList[2], Integer.parseInt(venueInfoList[3]));
+					db.insertVenue(venue);
+					// update the venue list table
+					TableItem temp = new TableItem(venueTable, SWT.None);
+					for (int i=0; i<venueInfoArray.length; i++){
+						temp.setText(i, venueInfoList[i]);
+					}
+				}
+			};
 			addVenuePage.pack();
 			addVenueShell.pack();
 			addVenueShell.open();
@@ -121,16 +146,43 @@ public class VenueManagement_VenueList extends Composite {
 	
 	public class delete extends SelectionAdapter {
 		public void widgetSelected(SelectionEvent e) {
-			if (venueTable.getColumnCount() != 0) {
+			if (venueTable.getItemCount() != 0) {
 				int index = venueTable.getSelectionIndex();
-				if (index < 0 || index >= venueTable.getItemCount()) {
-					// Do nothing.
-				} else {
-					/* update the itinerary table */
-					venueTable.remove(index);
-					/* update the database */
+				if (index >= 0 && index < venueTable.getItemCount()) {
 					DatabaseReader db = new DatabaseReader();
-					db.deleteVenue(db.getVenues().get(index));
+					ArrayList<Venue> venueList = db.getVenues();
+					Venue venueToBeDeleted = venueList.get(index);
+					ArrayList<VenueBookingApplication> vBookingAppList = db.getVenueBookingInfo(venueToBeDeleted);
+					if(!vBookingAppList.isEmpty()){
+						boolean deletable = true;
+						for (int i=0; i<vBookingAppList.size(); i++){
+							if(vBookingAppList.get(i).getDateTime().getDate().isNotEarlierThan(today) &&
+									vBookingAppList.get(i).getStatus() != MACRO.REJECTED){
+								deletable = false;
+								break;
+							}
+						}
+						if(deletable == false){
+							MessageBox warningPage  = new MessageBox(getDisplay().getActiveShell(), SWT.OK | SWT.ICON_WARNING );
+							warningPage.setText("Warning!");
+							warningPage.setMessage("There are still valid or unprocessed applications for this venue!");
+							warningPage.open(); 
+						}
+						else {
+							MessageBox warningPage  = new MessageBox(getDisplay().getActiveShell(), SWT.OK | SWT.ICON_WARNING );
+							warningPage.setText("Warning!");
+							warningPage.setMessage("Please clear its applicatin log before removing a venue!");
+							warningPage.open(); 
+						}
+					}
+					
+					else{
+						/* update the venue list table */
+						venueTable.remove(index);
+						/* update the database */
+						db.deleteVenue(db.getVenues().get(index));
+						
+					}
 				}
 			}
 		}
@@ -138,10 +190,40 @@ public class VenueManagement_VenueList extends Composite {
 	
 	public class edit extends SelectionAdapter {
 		public void widgetSelected(SelectionEvent e) {
-			int index = venueTable.getSelectionIndex();
+			final int index = venueTable.getSelectionIndex();
 			if(index >=0 && index < venueTable.getItemCount()){
 				Shell editVenueShell = new Shell(getDisplay());
-				VenueListEditItem editVenuePage = new VenueListEditItem(editVenueShell, SWT.None, venueTable, index);
+				Image icon = new Image(getDisplay(), "resources/eManager.png");
+				editVenueShell.setText("eManager - Edit a venue");
+				editVenueShell.setImage(icon);
+				AbstractEdit editVenuePage = new AbstractEdit(editVenueShell, SWT.None, 
+						venueInfoArray, venueInfoSignatureArray, venueTable){
+					// get data from table selection
+					public void onLoad(){
+						for (int i=0; i<venueInfoArray.length; i++){
+							setData(venueTable.getItem(index).getText(i), 
+									venueInfoSignatureArray[i], i);
+						}
+					}
+					// submit modified info
+					public void onSubmit(){
+						String[] venueInfoList = getStringList();
+						// find the venue to be edited
+						Venue venue = db.getVenues().get(index);
+						venue.setName(venueInfoList[0]);
+						venue.setLocation(venueInfoList[1]);
+						venue.setType(venueInfoList[2]);
+						venue.setCapacity(Integer.parseInt(venueInfoList[3]));
+						// update the database
+						db.updateVenue(venue);
+						// update the venue list table
+						TableItem temp = new TableItem(venueTable, SWT.None);
+						for (int i=0; i<venueInfoArray.length; i++){
+							temp.setText(i, venueInfoList[i]);
+						}	
+					}
+				};
+				
 				editVenuePage.pack();
 				editVenueShell.pack();
 				editVenueShell.open();
